fix(planner): Corregir interpretación de mold_quantity como fracción de caja
**Problema identificado:**
- mold_quantity NO es número entero de moldes
- mold_quantity = 1/piezas_por_molde (fracción de caja que usa UNA pieza)
- Ejemplo: si piezas_por_molde=4, entonces mold_quantity=0.25

**Cambios realizados:**

1. Schema (data_schema.py):
   - Documentar que mold_quantity debe ser REAL para fracciones
   - SQLite INTEGER acepta REAL, no requiere migración forzada

2. Importación (data_repository.py):
   - Cambiar lógica 'or 1' que convertía 0.0 en 1.0
   - Validar explícitamente: si <=0 entonces default 1.0
   - Agregar comentario explicativo sobre semántica

3. Actualización recursos diarios (planner_repository.py):
   - update_daily_resources_from_demolding(): acumular fracciones por día/flask_type
   - Usar math.ceil() para redondear hacia arriba al actualizar disponibilidad
   - Ejemplo: 3 piezas  0.25 = 0.75 cajas  ceil(0.75) = 1 caja ocupada

4. get_flasks_in_use_from_demolding():
   - Cambiar de int() a float() para mold_quantity
   - Acumular fracciones en dict auxiliar
   - Aplicar ceil() al retornar conteos enteros

5. Tests y diagnóstico:
   - Script test_mold_qty_migration.py para verificar schema
   - Verificar que valores existentes se procesen correctamente

**Resultado esperado:**
Al importar snapshot desmoldeo con valores decimales (ej: 0.25, 0.5),
la tabla de condiciones iniciales mostrará correctamente las cajas ocupadas
sumando todas las fracciones de piezas en enfriamiento.

+++ src/foundryplan/data/schema/data_schema.py
@@ -165,7 +165,7 @@ def ensure_schema(con: sqlite3.Connection) -> None:
             poured_date TEXT,
             poured_time TEXT,
             cooling_hours REAL,
-            mold_quantity INTEGER
+            mold_quantity REAL
         );
 
         CREATE TABLE IF NOT EXISTS orders (
@@ -292,3 +292,7 @@ def ensure_schema(con: sqlite3.Connection) -> None:
         con.execute("ALTER TABLE sap_demolding_snapshot ADD COLUMN cancha TEXT")
     except Exception:
         pass
+    
+    # Note: mold_quantity should be REAL to store fractions (1/piezas_por_molde)
+    # SQLite's INTEGER affinity can store REAL values, but for new tables we use REAL
+    # Existing data will work correctly with float() conversion in Python
